requirements.txt
ccxt
pandas
numpy
python-telegram-bot==20.0a6
python-dotenv
